requirements.txt
Django==5.1.4
djangorestframework==3.16.0
django-filter==25.1